feat(feature/donkey/#35): 게시글 상세페이지 컴포넌트 분리
